requirements.txt
python-telegram-bot==20.7
beautifulsoup4==4.12.3
playwright==1.41.2
asyncio